requirements.txt
numpy
pandas
python-docx
torch>=1.4
configparser
xlrd
openpyxl
tensorboardX
sklearn
PyYAML
h5py
torch_optimizer
matplotlib
kaggle
black
flake8
tensorflow